+++ src/prepared_statement.rs
@@ -4,7 +4,7 @@ use crate::protocol::part::Part;
 use crate::protocol::partkind::PartKind;
 use crate::protocol::parts::hdb_value::HdbValue;
 use crate::protocol::parts::lob_flags::LobFlags;
-use crate::protocol::parts::parameter_descriptor::{ParameterDescriptor, ParameterDescriptors};
+use crate::protocol::parts::parameter_descriptor::ParameterDescriptors;
 use crate::protocol::parts::parameter_rows::ParameterRows;
 use crate::protocol::parts::resultset_metadata::ResultSetMetadata;
 use crate::protocol::parts::type_id::TypeId;
@@ -26,18 +26,18 @@ use std::sync::{Arc, Mutex};
 ///
 /// ## Type systems
 ///
-/// It is helpful to understand the involved type systems - there are four!
+/// There are four typesystems involved!
 /// * The _database type system_ consists of the standard SQL types and proprietary types
 ///   to represent values,
 ///   like TINYINT, FLOAT, NVARCHAR, and many others.
-///   This type system is NOT directly visible to the client!
+///   This type system is NOT directly visible to the client.
 /// * The _wire_ has its own type system - it's focus is on efficient data transfer.
-///   hdbconnect has to deal with these types appropriately.
+///   hdbconnect has to deal with these types internally.
 ///   `hdbconnect::TypeId` enumerates a somewhat reduced superset of the server-side
 ///   and the wire type system.
-/// * The _driver API_ represents values as variants of `HdbValue`; this type system
+/// * The _driver API_ represents values with the `enum` `HdbValue`; this type system
 ///   hides the complexity of the wire type system and aims to be as close to the rust type system
-///   as possible
+///   as possible.
 /// * The application is written in rust, and uses the _rust type system_.
 ///
 /// ## From Rust to HdbValue
@@ -71,7 +71,7 @@ use std::sync::{Arc, Mutex};
 pub struct PreparedStatement {
     am_ps_core: Arc<Mutex<PreparedStatementCore>>,
     server_usage: ServerUsage,
-    o_a_descriptors: Option<Arc<ParameterDescriptors>>,
+    a_descriptors: Arc<ParameterDescriptors>,
     o_a_rsmd: Option<Arc<ResultSetMetadata>>,
     batch: ParameterRows<'static>,
     _o_table_location: Option<Vec<i32>>,
@@ -133,12 +133,10 @@ impl<'a> PreparedStatement {
     /// ```
     pub fn execute<T: serde::ser::Serialize>(&mut self, input: &T) -> HdbResult<HdbResponse> {
         trace!("PreparedStatement::execute()");
-        if let Some(ref a_descriptors) = self.o_a_descriptors {
-            if a_descriptors.has_in() {
-                let mut par_rows = ParameterRows::new();
-                par_rows.push(input, &a_descriptors)?;
-                return self.execute_parameter_rows(Some(par_rows));
-            }
+        if self.a_descriptors.has_in() {
+            let mut par_rows = ParameterRows::new();
+            par_rows.push(input, &self.a_descriptors)?;
+            return self.execute_parameter_rows(Some(par_rows));
         }
         self.execute_parameter_rows(None)
     }
@@ -188,92 +186,88 @@ impl<'a> PreparedStatement {
     /// After this call, the data from the readers are transferred to the database in additional
     /// roundtrips.
     pub fn execute_row(&'a mut self, hdb_values: Vec<HdbValue<'a>>) -> HdbResult<HdbResponse> {
-        if let Some(ref descriptors) = self.o_a_descriptors {
-            if descriptors.has_in() {
-                let mut request = Request::new(RequestType::Execute, HOLD_CURSORS_OVER_COMMIT);
+        if self.a_descriptors.has_in() {
+            let mut request = Request::new(RequestType::Execute, HOLD_CURSORS_OVER_COMMIT);
 
-                request.push(Part::new(
-                    PartKind::StatementId,
-                    Argument::StatementId(self.am_ps_core.lock()?.statement_id),
-                ));
+            request.push(Part::new(
+                PartKind::StatementId,
+                Argument::StatementId(self.am_ps_core.lock()?.statement_id),
+            ));
 
-                // If readers were provided, pick them out and replace them with None
-                let mut readers: Vec<(HdbValue, TypeId)> = vec![];
-                let hdb_values = hdb_values
-                    .into_iter()
-                    .zip(descriptors.iter_in())
-                    .map(|(v, d)| {
-                        if let HdbValue::LOBSTREAM(Some(_)) = v {
-                            readers.push((v, d.type_id()));
-                            HdbValue::LOBSTREAM(None)
-                        } else {
-                            v
-                        }
-                    })
-                    .collect();
-
-                let mut par_rows = ParameterRows::new();
-                par_rows.push_hdb_values(hdb_values, &descriptors)?;
-                request.push(Part::new(
-                    PartKind::Parameters,
-                    Argument::Parameters(par_rows),
-                ));
-
-                if self
-                    .am_ps_core
-                    .lock()?
-                    .am_conn_core
-                    .lock()?
-                    .connect_options()
-                    .get_implicit_lob_streaming()
-                    .unwrap_or(false)
-                {
-                    request.push(Part::new(
-                        PartKind::LobFlags,
-                        Argument::LobFlags(LobFlags::for_implicit_streaming()),
-                    ));
-                }
-
-                let mut main_reply = self.am_ps_core.lock()?.am_conn_core.full_send(
-                    request,
-                    self.o_a_rsmd.clone(),
-                    self.o_a_descriptors.clone(),
-                    &mut None,
-                )?;
-
-                if let Some(Argument::WriteLobReply(wlr)) =
-                    main_reply.extract_first_arg_of_type(PartKind::WriteLobReply)
-                {
-                    let locator_ids = wlr.into_locator_ids();
-                    if locator_ids.len() != readers.len() {
-                        return Err(HdbError::Usage(format!(
-                            "The number of provided readers ({}) does not match \
-                             the number of required readers ({})",
-                            readers.len(),
-                            locator_ids.len()
-                        )));
+            // If readers were provided, pick them out and replace them with None
+            let mut readers: Vec<(HdbValue, TypeId)> = vec![];
+            let hdb_values = hdb_values
+                .into_iter()
+                .zip(self.a_descriptors.iter_in())
+                .map(|(v, d)| {
+                    if let HdbValue::LOBSTREAM(Some(_)) = v {
+                        readers.push((v, d.type_id()));
+                        HdbValue::LOBSTREAM(None)
+                    } else {
+                        v
                     }
-                    for (locator_id, (reader, type_id)) in locator_ids.into_iter().zip(readers) {
-                        debug!("writing content to locator with id {:?}", locator_id);
-                        if let HdbValue::LOBSTREAM(Some(reader)) = reader {
-                            let mut reader = reader.lock().unwrap();
-                            let mut writer = LobWriter::new(
-                                locator_id,
-                                type_id,
-                                self.am_ps_core.lock()?.am_conn_core.clone(),
-                            )?;
-                            std::io::copy(&mut *reader, &mut writer)?;
-                            writer.flush()?;
-                        }
-                    }
-                }
-                main_reply.into_hdbresponse(
-                    &mut (self.am_ps_core.lock()?.am_conn_core),
-                    Some(&mut self.server_usage),
-                )
-            } else {
-                self.execute_parameter_rows(None)
+                })
+                .collect();
+
+            let mut par_rows = ParameterRows::new();
+            par_rows.push_hdb_values(hdb_values, &self.a_descriptors)?;
+            request.push(Part::new(
+                PartKind::Parameters,
+                Argument::Parameters(par_rows),
+            ));
+
+            if self
+                .am_ps_core
+                .lock()?
+                .am_conn_core
+                .lock()?
+                .connect_options()
+                .get_implicit_lob_streaming()
+                .unwrap_or(false)
+            {
+                request.push(Part::new(
+                    PartKind::LobFlags,
+                    Argument::LobFlags(LobFlags::for_implicit_streaming()),
+                ));
             }
+
+            let mut main_reply = self.am_ps_core.lock()?.am_conn_core.full_send(
+                request,
+                self.o_a_rsmd.clone(),
+                Some(self.a_descriptors.clone()),
+                &mut None,
+            )?;
+
+            if let Some(Argument::WriteLobReply(wlr)) =
+                main_reply.extract_first_arg_of_type(PartKind::WriteLobReply)
+            {
+                let locator_ids = wlr.into_locator_ids();
+                if locator_ids.len() != readers.len() {
+                    return Err(HdbError::Usage(format!(
+                        "The number of provided readers ({}) does not match \
+                         the number of required readers ({})",
+                        readers.len(),
+                        locator_ids.len()
+                    )));
+                }
+                for (locator_id, (reader, type_id)) in locator_ids.into_iter().zip(readers) {
+                    debug!("writing content to locator with id {:?}", locator_id);
+                    if let HdbValue::LOBSTREAM(Some(reader)) = reader {
+                        let mut reader = reader.lock().unwrap();
+                        let mut writer = LobWriter::new(
+                            locator_id,
+                            type_id,
+                            self.am_ps_core.lock()?.am_conn_core.clone(),
+                        )?;
+                        std::io::copy(&mut *reader, &mut writer)?;
+                        writer.flush()?;
+                    }
+                }
+            }
+            main_reply.into_hdbresponse(
+                &mut (self.am_ps_core.lock()?.am_conn_core),
+                Some(&mut self.server_usage),
+            )
         } else {
             self.execute_parameter_rows(None)
         }
@@ -283,11 +277,9 @@ impl<'a> PreparedStatement {
     /// `PreparedStatement`, if it is consistent with the metadata.
     pub fn add_batch<T: serde::ser::Serialize>(&mut self, input: &T) -> HdbResult<()> {
         trace!("PreparedStatement::add_batch()");
-        if let Some(ref descriptors) = self.o_a_descriptors {
-            if descriptors.has_in() {
-                self.batch.push(input, &descriptors)?;
-                return Ok(());
-            }
+        if self.a_descriptors.has_in() {
+            self.batch.push(input, &self.a_descriptors)?;
+            return Ok(());
         }
         Err(HdbError::Serialization(
             SerializationError::StructuralMismatch("no metadata in add_batch()"),
@@ -302,11 +294,10 @@ impl<'a> PreparedStatement {
     /// Note that LOB streaming can not be combined with using the batch.
     pub fn add_row_to_batch(&mut self, hdb_values: Vec<HdbValue<'static>>) -> HdbResult<()> {
         trace!("PreparedStatement::add_row_to_batch()");
-        if let Some(ref descriptors) = self.o_a_descriptors {
-            if descriptors.has_in() {
-                self.batch.push_hdb_values(hdb_values, &descriptors)?;
-                return Ok(());
-            }
+        if self.a_descriptors.has_in() {
+            self.batch
+                .push_hdb_values(hdb_values, &self.a_descriptors)?;
+            return Ok(());
         }
         Err(HdbError::Serialization(
             SerializationError::StructuralMismatch("no metadata in add_row_to_batch()"),
@@ -319,25 +310,19 @@ impl<'a> PreparedStatement {
     /// If the statement does not need input and the batch is empty,
     /// a single execution is triggered.
     pub fn execute_batch(&mut self) -> HdbResult<HdbResponse> {
-        if self.batch.is_empty() {
-            if let Some(ref descriptors) = self.o_a_descriptors {
-                if descriptors.has_in() {
-                    return Err(HdbError::Usage(
-                        "The batch is empty and cannot be executed".to_string(),
-                    ));
-                }
-            }
+        if self.batch.is_empty() && self.a_descriptors.has_in() {
+            return Err(HdbError::Usage(
+                "The batch is empty and cannot be executed".to_string(),
+            ));
         }
         let mut rows2 = ParameterRows::new();
         mem::swap(&mut self.batch, &mut rows2);
         self.execute_parameter_rows(Some(rows2))
     }
 
-    /// Descriptors of all parameters of the prepared statement (in, out, inout), if any.
-    pub fn parameter_descriptors(&self) -> Option<&[ParameterDescriptor]> {
-        self.o_a_descriptors
-            .as_ref()
-            .map(|descr| descr.ref_inner().as_slice())
+    /// Descriptors of all parameters of the prepared statement (in, out, inout).
+    pub fn parameter_descriptors(&self) -> Arc<ParameterDescriptors> {
+        Arc::clone(&self.a_descriptors)
     }
 
     fn execute_parameter_rows(&mut self, o_rows: Option<ParameterRows>) -> HdbResult<HdbResponse> {
@@ -354,7 +339,7 @@ impl<'a> PreparedStatement {
         let reply = self.am_ps_core.lock()?.am_conn_core.full_send(
             request,
             self.o_a_rsmd.clone(),
-            self.o_a_descriptors.clone(),
+            Some(self.a_descriptors.clone()),
             &mut None,
         )?;
         let mut response = reply.into_hdbresponse(
@@ -386,14 +371,14 @@ impl<'a> PreparedStatement {
         // TableLocation, TransactionFlags,
         let mut o_table_location: Option<Vec<i32>> = None;
         let mut o_stmt_id: Option<u64> = None;
-        let mut o_a_descriptors: Option<Arc<ParameterDescriptors>> = None;
+        let mut a_descriptors: Arc<ParameterDescriptors> = Arc::new(ParameterDescriptors::new());
         let mut o_a_rsmd: Option<Arc<ResultSetMetadata>> = None;
         let mut server_usage: ServerUsage = Default::default();
 
         while !reply.parts.is_empty() {
             match reply.parts.pop_arg() {
                 Some(Argument::ParameterMetadata(descriptors)) => {
-                    o_a_descriptors = Some(Arc::new(descriptors));
+                    a_descriptors = Arc::new(descriptors);
                 }
                 Some(Argument::StatementId(id)) => {
                     o_stmt_id = Some(id);
@@ -433,7 +418,7 @@ impl<'a> PreparedStatement {
 
         debug!(
             "PreparedStatement created with parameter descriptors = {:?}",
-            o_a_descriptors
+            a_descriptors
         );
 
         let am_ps_core = Arc::new(Mutex::new(PreparedStatementCore {
@@ -444,7 +429,7 @@ impl<'a> PreparedStatement {
             am_ps_core,
             server_usage,
             batch: ParameterRows::new(),
-            o_a_descriptors,
+            a_descriptors,
             o_a_rsmd,
             _o_table_location: o_table_location,
         })