+++ tests/test_028_timestamp.rs
@@ -73,11 +73,7 @@ fn test_timestamp(
             .prepare("select sum(number) from TEST_TIMESTAMP where mydate = ? or mydate = ?")?;
         // Enforce that NaiveDateTime values are converted in the client (with serde) to the DB type:
         prep_stmt.add_batch(&(naive_datetime_values[2], naive_datetime_values[3]))?;
-        let mut response = prep_stmt.execute_batch()?;
-        let pds = response.get_parameter_descriptors()?;
-        debug!("1st Parameter Descriptor: {:?}", pds[0]);
-        debug!("2nd Parameter Descriptor: {:?}", pds[1]);
-        assert_eq!(pds.len(), 2);
+        let response = prep_stmt.execute_batch()?;
         let typed_result: i32 = response.into_resultset()?.try_into()?;
         assert_eq!(typed_result, 31);